pyproject.toml
[project]
name = "ts-pit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.25.1",
    "fastapi>=0.128.0",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "yfinance>=1.1.0",
    "colorama>=0.4.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "beautifulsoup4>=4.14.3",
    "requests>=2.32.5",
    "ddgs>=9.10.0",
    "aiohttp>=3.13.3",
    "faker>=40.1.2",
    "loguru>=0.7.3",
]
